139 fix minor logic improvements

+++ src/main/java/com/salesianostriana/dam/ecoshop/controller/CustomerController.java
@@ -1,8 +1,11 @@
 package com.salesianostriana.dam.ecoshop.controller;
 
+import java.security.Principal;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -42,8 +45,8 @@ public class CustomerController {
 	*/
 	
 	@GetMapping({"/", "/list"})
-	public String findAll (Model model) {
-		model.addAttribute("customers", service.findAll());
+	public String findAll (Model model, Principal principal) {
+		model.addAttribute("customers", service.findAllExceptUsername(principal.getName()));
 		return "customers/list";
 	}
 	
@@ -79,14 +82,23 @@ public class CustomerController {
 	
 	@PreAuthorize("hasRole('ADMIN')")
 	@GetMapping("/delete/{id}")
-	public String delete (@PathVariable Long id, Model model) {
+	public String delete (@PathVariable Long id, Authentication authentication, Model model) {
 		
 		Optional<Customer> customer = service.findById(id);
 		
-		if(customer.isPresent()) 
-			service.delete(customer.get());
-		
-		return "redirect:/customers/list";
+		//aunque no sale en la lista de clientes, también lo podría hacer con la ruta, por lo que necesita esta 'validación' para que no haga /customers/delete/1, siendo 1 el admin y estando logueado en esa acc
+		if (customer.isPresent()) {
+	        String currentUsername = authentication.getName();
+
+	        if (customer.get().getUser() != null && customer.get().getUser().getUsername().equals(currentUsername)) {
+	            return "redirect:/customers/list";
+	        }
+
+	        service.delete(customer.get());
+	    }
+	    return "redirect:/customers/list";
 	}
+	
+	
 	
 }
